+++ lib/eve_online.rb
@@ -51,7 +51,9 @@ require 'eve_online/esi/corporation'
 require 'eve_online/esi/corporation_industry_jobs'
 require 'eve_online/esi/corporation_blueprints'
 
+require 'eve_online/esi/alliances'
 require 'eve_online/esi/alliance'
+require 'eve_online/esi/alliance_corporations'
 require 'eve_online/esi/alliance_icon'
 
 require 'eve_online/esi/dogma_attributes'